added a base abstraction and divided the retrieval module into two
